+++ tests/com/tkomiya/testutils/TestTranslationLoader.java
@@ -4,10 +4,6 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 public class TestTranslationLoader {
-
-  public static JSONObject mondayEnglishJson() throws JSONException {
-    return TestTranslationLoaderUtil.makeTranslationJson("Monday", 0, 0, "");
-  }
   
   public static JSONObject mondayKoreanJson() throws JSONException {
     return TestTranslationLoaderUtil.makeTranslationJson("월요일", 1, 1, "");
@@ -17,10 +13,6 @@ public class TestTranslationLoader {
     return TestTranslationLoaderUtil.makeTranslationJson("月曜日", 2, 2, "");
   }
   
-  public static JSONObject tuesdayEnglishJson() throws JSONException {
-    return TestTranslationLoaderUtil.makeTranslationJson("Tuesday", 4, 5, "1991-07-17T06:40:44.000Z");
-  }
-  
   public static JSONObject tuesdayKoreanJson() throws JSONException {
     return TestTranslationLoaderUtil.makeTranslationJson("화요일", 6, 14, "1991-07-17T06:40:44.000Z");
   }
@@ -28,10 +20,7 @@ public class TestTranslationLoader {
   public static JSONObject tuesdayJapaneseJson() throws JSONException {
     return TestTranslationLoaderUtil.makeTranslationJson("火曜日", 1, 7, "1991-07-17T06:40:44.000Z");  
   }
-  
-  public static JSONObject wednesdayEnglishJson() throws JSONException {
-    return TestTranslationLoaderUtil.makeTranslationJson("Wednesday", 0, 0, "2015-01-15T12:15:44.000Z");
-  }
+
   
   public static JSONObject wednesdayKoreanJson() throws JSONException {
     return TestTranslationLoaderUtil.makeTranslationJson("수요일", 1, 7, "2015-01-15T12:15:44.000Z");
